+++ test/tarray.c
@@ -507,7 +507,7 @@ test_array_compound_atomic(void)
     for(i=0; i<SPACE1_DIM1; i++)
         for(j=0; j<ARRAY1_DIM1; j++) {
             wdata[i][j].i=i*10+j;
-            wdata[i][j].f=i*2.5+j;
+            wdata[i][j].f=(float)(i*2.5+j);
         } /* end for */
 
     /* Create file */
@@ -724,7 +724,7 @@ test_array_compound_array(void)
         for(j=0; j<ARRAY1_DIM1; j++) {
             wdata[i][j].i=i*10+j;
             for(k=0; k<ARRAY1_DIM1; k++)
-                wdata[i][j].f[k]=i*10+j*2.5+k;
+                wdata[i][j].f[k]=(float)(i*10+j*2.5+k);
         } /* end for */
 
     /* Create file */
@@ -1536,7 +1536,7 @@ test_array_bkg(void)
 		for (j = 0; j < ALEN; j++)
 		  {
 			cf[i].a[j] = 100*(i+1) + j;
-			cf[i].b[j] = 100.*(i+1) + 0.01*j;
+			cf[i].b[j] = (float)(100.*(i+1) + 0.01*j);
 			cf[i].c[j] = 100.*(i+1) + 0.02*j;
 		  }
 	  }
@@ -1679,7 +1679,7 @@ test_array_bkg(void)
     /* -------------------------------- */
     for (i=0; i< LENGTH; i++)
         for (j = 0; j < ALEN; j++)
-            cf[i].b[j]=fld[i].b[j] = 1.313;
+            cf[i].b[j]=fld[i].b[j] = (float)1.313;
 
     status = H5Dwrite (dataset, type, H5S_ALL, H5S_ALL, H5P_DEFAULT, fld);
     CHECK(status, FAIL, "H5Dwrite");
